refactor: move DetailsCard to src/components dir
