+++ src/components/DetailsCard.jsx
